Fix chrome driver, add Page Object, elements and first operations
