+++ .gitignore
@@ -3,3 +3,5 @@
 .classpath
 .project
 /bin/
+.idea
+.allure
