+++ src/main/java/com/sample/test/pizzaForm/constants/PizzaToppings.java
@@ -1,0 +1,29 @@
+package com.sample.test.pizzaForm.constants;
+
+
+public enum PizzaToppings {
+    MANGOS("Diced Mango"),
+    OLIVES("Olives"),
+    MUSHROOMS("Mushrooms"),
+    ONIONS("Caramelized onions"),
+    ITALIANHAM("Italian Ham"),
+    PEPPERONI("Classic Pepperoni"),
+    SALAMI("Salami"),
+    PROVOLNE("Provolone cheese"),
+    EXTRACHEESE("Extra cheese"),
+    MOZZARELLA("Mozzarella cheese"),
+    PARMASAN("Parmesan cheese")
+
+    ;
+
+    private String displayName;
+
+    private PizzaToppings(String displayName) {
+        this.displayName = displayName;
+    }
+
+    public String getDisplayName() {
+        return displayName;
+    }
+
+}
