--- src/test/java/com/sample/test/demo/tests/DemoTest.java
@@ -1,13 +1,0 @@
-package com.sample.test.demo.tests;
-
-import org.testng.annotations.Test;
-import com.sample.test.demo.TestBase;
-
-public class DemoTest extends TestBase {
-
-    @Test
-    public void demoTest() {
-        System.out.println("HELLO WORLD");
-    }
-
-}
