+++ src/test/resources/config.properties
@@ -1,5 +1,5 @@
 browser=chrome
-platform=mac
-url=http\://wwww.google.com
+platform=windows
+url=file:///C:/Users/Ihor_Hryhorenko/IdeaProjects/SQEDemoChallengeUI/src/test/resources/files/index.html
